Rename the 1.47 ABI to 1.48
